[optimize]優化friendType
