+++ types/api.ts
@@ -1,5 +1,7 @@
+/** 通用 API Response Type */
 export type apiResponseType = {
   code: string;
+  message: string;
   data: unknown;
   error: unknown;
 };
