+++ types/userType.ts
@@ -27,22 +27,3 @@ export type userSettingsType = {
   themeMode: string;
   language: string;
 };
-
-/** 好友狀態資料 */
-export type friendStatusDataType = {
-  uid: string;
-  status: number;
-  createdAt: any;
-  sourceUserData: userDataType;
-};
-
-/** 好友資料 */
-export type friendDataType = {
-  uid: string;
-  userName: string;
-  avatarUrl: string;
-  bgColor: string;
-  status: number;
-  createdAt: string;
-  sourceUserData: userDataType;
-};
